Karolayn

+++ src/main/java/com/accesadades/botiga/Service/CategoriaServiceImpl.java
@@ -2,11 +2,14 @@ package com.accesadades.botiga.Service;
 
 import com.accesadades.botiga.Repository.CategoriaRepository;
 import com.accesadades.botiga.Model.Categoria;
+import com.accesadades.botiga.DTO.CategoriaDTO;
 import com.accesadades.botiga.Mapper.BotigaMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import java.util.List;
+
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class CategoriaServiceImpl implements CategoriaService {
@@ -15,30 +18,50 @@ public class CategoriaServiceImpl implements CategoriaService {
     private CategoriaRepository categoriaRepository;
 
     @Autowired
-    private BotigaMapper categoriaMapper;
+    private BotigaMapper mapper;
 
     @Override
-    public List<Categoria> findAll() {
-        return categoriaRepository.findAll();
+    public CategoriaDTO save(CategoriaDTO categoriaDTO) {
+        Categoria categoria = mapper.toEntity(categoriaDTO);
+        return mapper.toDTO(categoriaRepository.save(categoria));
     }
 
     @Override
-    public Optional<Categoria> findById(Long id) {
-        return categoriaRepository.findById(id);
+    public Set<CategoriaDTO> findAll() {
+        return categoriaRepository.findAll()
+                .stream()
+                .map(mapper::toDTO) 
+                .collect(Collectors.toSet());
     }
 
     @Override
-    public Categoria save(Categoria entity) {
-        return categoriaRepository.save(entity);
+    public Optional<CategoriaDTO> findById(Long id) {
+        return categoriaRepository.findById(id)
+                .map(mapper::toDTO);
     }
 
     @Override
-    public void deleteById(Long id) {
-        categoriaRepository.deleteById(id);
+    public Optional<CategoriaDTO> findByDescCategoria(String desc) {
+        return categoriaRepository.findByDescCategoria(desc)
+                .map(mapper::toDTO);
     }
 
     @Override
-    public Optional<Categoria> findByDescCategoria(String desc) {
-        return categoriaRepository.findByDescCategoria(desc);
+    public boolean delete(Long id) {
+        if (categoriaRepository.existsById(id)) {
+            categoriaRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean update(CategoriaDTO categoriaDTO) {
+        if (categoriaDTO.getId() == null || !categoriaRepository.existsById(categoriaDTO.getId())) {
+            return false;
+        }
+        Categoria updatedCategoria = mapper.toEntity(categoriaDTO);
+        categoriaRepository.save(updatedCategoria);
+        return true;
     }
 }
